requirements-gpu.txt
tensorflow-gpu
keras
setuptools
matplotlib
numpy
pandas
jupyter
scikit-learn
h5py
requests
cherrypy